hotfix: #14 - Double $nextTick required for DOM search
i'm not certain why we need the double nextTick before handling open, but the markup version of this plugin already had it before, so this was regression.

I suspect that might be related with the mount process of portal, but requires more investigation. I'll ship this hotfix as this breaks dialogs that start open.