fix: do not close kedown.esc, if focused on search input and not empty
default behaviour of search inputs is to reset input value on escape, so we don't wan't to interfere with that.

If we really want want to prevent close on this or any input, just need to stopPropagation on keydown events `keydow.prevent`. This wills top bubbling and prevent dialog from closing