fix: #21 handling focus mutations with conditionals
Some actions might trigger conditionals that change focus references within the dialog.
Some could be added some removed. We were already updating those in the mutation observer (now
updated with style changes as well for v-show), but if the focused element was removed itself
focus was lost and the dialog could be closed with escape key.

Among other problems this now moves focus to the next focusable element within the dialog
or first it it doesn't exists. That guarantes that event if activeElement disappears, focus is kept
trapped inside dialog.

update: listen to focus to determine focuswed element in the focus tree

this was done on trapFocus(), but it would only trigger the current if we actually tab our way there. If we click a button we didn't knew what was his position in current focusable tree.

We now save it as instance property

update: move focus to next element within the dialog if current disappears

- after an async action the trigger might be removed or hidden, in both cases we move focus to the next element, it doesn't exist we go back to first.

This fixes #21