fix(focus-trap): only querySelector valid autofocus elements
avoids moving initial focus to invalid <div autofocus> which aren't focusable anyways. If people want invalid els to be focusable, they can pass it via focusRef (by binding it to an invalid element) or via focusTrapCreateOptions object, but by default we won't do it because it promotes a bad practice.