chore(deps): add @linusborg/vue-simple-portal as optional dependency
appending to body is not part of wai-aria guidelines so it's not mandatory to implement as core of this Component. That being said, we couldn't recomend enough the usage of a portal solution, and this simple version of @linusborg is a perfect fit for vue2 apps (teleport equivalent for vue3). Regular portal-vue can still be used, but unless you're already using it for other ppurposes, it's an overkill for the dialog needs