fix!: add required rootRef, removed pseudo-drag to outside closing workaround
The pseudo-drag workaround was done because our bakdropRef was being applied to the root element of dialog. I've used some mousedown/up trickery + e.stopPropagtion to fix it. e.stopPropagation is never a good id (see clickOutside repos). Any element rendered inside the dialog that relied on event bubbling (default behaviour) would now be broken: click outside of dropdowns inside dialog would not work and they would never be closed.

By using a root element, and ensure backdrop is an optional and separate element we not longer need the hack to prevent "drag outside" neither stopPropagation which is win, even if it's a breaking change.

BREAKING-CHANGE:
- Your dialog compositions now must have the`rootRef` binding as the root element of your dialog.
-Revert release https://github.com/renatodeleao/a11y-vue-dialog/releases/tag/v0.4.4